Try to split faces into subfaces
Currently the code will not split, but rather clone each face based on the number of divisions. E.g. one division is splitting the x and y axis in half, so should result in 4 subfaces per face. This current iteration of the code does not generate 4 adjacent subfaces yet, but rather 4 identical copies of the original face.