Add list of tasks
These get passed into a LoadTask object by AddTasks(), where the function that is to be called gets made into a System.Action by passing it as the body of a lambda function. Enabled the oceanLoader.Load() task, and wrote an empty Load method for that class so it could suppress the compiler warnings.